@@ -15,14 +15,14 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.1.RELEASE</version>
+		<version>2.7.18</version>
 		<relativePath />
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
+		<java.version>21</java.version>
 		<shiro.version>1.11.0</shiro.version>
 		<thymeleaf.extras.shiro.version>2.0.0</thymeleaf.extras.shiro.version>
 		<mybatis.spring.boot.starter.version>1.3.2</mybatis.spring.boot.starter.version>
@@ -87,8 +87,8 @@
 
 		<!-- Mysql驱动包 -->
 		<dependency>
-			<groupId>mysql</groupId>
-			<artifactId>mysql-connector-java</artifactId>
+			<groupId>com.mysql</groupId>
+			<artifactId>mysql-connector-j</artifactId>
 		</dependency>
 
 		<!-- SpringBoot集成mybatis框架 -->
@@ -198,6 +198,11 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
+  <dependency>
+   <groupId>org.springdoc</groupId>
+   <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+   <version>1.5.x</version>
+  </dependency>
 
 		<!--velocity代码生成使用模板 -->
 		<dependency>
@@ -220,18 +225,12 @@
 		</dependency>
 
 		<!-- swagger2 -->
-		<dependency>
-			<groupId>io.springfox</groupId>
-			<artifactId>springfox-swagger2</artifactId>
-			<version>${swagger.version}</version>
-		</dependency>
+  <dependency>
+   <groupId>jakarta.annotation</groupId>
+   <artifactId>jakarta.annotation-api</artifactId>
+  </dependency>
 
 		<!-- swagger2-UI -->
-		<dependency>
-			<groupId>io.springfox</groupId>
-			<artifactId>springfox-swagger-ui</artifactId>
-			<version>${swagger.version}</version>
-		</dependency>
 
 		<!-- 获取系统信息 -->
 		<dependency>
@@ -300,6 +299,16 @@
 				<artifactId>netty-all</artifactId>
 				<version>5.0.0.Alpha2</version>
 			</dependency>
+   <dependency>
+    <groupId>net.java.dev.jna</groupId>
+    <artifactId>jna</artifactId>
+    <version>4.5.2</version>
+   </dependency>
+   <dependency>
+    <groupId>net.java.dev.jna</groupId>
+    <artifactId>jna-platform</artifactId>
+    <version>4.5.2</version>
+   </dependency>
 		</dependencies>
 	</dependencyManagement>
 
@@ -349,8 +358,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+     <release>${java.version}</release>
                 </configuration>
             </plugin>
 
