@@ -144,7 +144,7 @@ public class TaskExecuteServiceImpl implements ITaskExecuteService
 			}
 			
 			if(!PermissionUtils.isProjectPermsPassByProjectId(taskExecute.getProjectId())){	
-				  throw new BusinessException(String.format("执行任务【%1$s】没有项目删除权限", taskExecute.getTaskName()));
+				  throw new BusinessException("执行任务【%1$s】没有项目删除权限".formatted(taskExecute.getTaskName()));
 			}
 			
 			//30分钟内的任务不允许删除，防止客户端提交脏数据
@@ -152,7 +152,7 @@ public class TaskExecuteServiceImpl implements ITaskExecuteService
 			boolean isTaskDeleteTime = (0==taskExecute.getTaskStatus() || 1==taskExecute.getTaskStatus())
 					&& taskExecute.getUpdateTime().after(date);			
 			if (isTaskDeleteTime) {
-				throw new BusinessException(String.format("执行中的任务 %1$s 必须30分钟后才可以删除", taskExecute.getTaskName()));
+				throw new BusinessException("执行中的任务 %1$s 必须30分钟后才可以删除".formatted(taskExecute.getTaskName()));
 			}
 			
 			taskCaseLogMapper.deleteTaskCaseLogByTaskId(taskId);
