@@ -2,6 +2,9 @@ package com.luckyframe.project.tool.swagger;
 
 import java.util.ArrayList;
 import java.util.List;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,16 +13,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.luckyframe.framework.web.controller.BaseController;
 import com.luckyframe.framework.web.domain.AjaxResult;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiOperation;
 
 /**
  * swagger 测试方法
  * 
  * @author ruoyi
  */
-@Api("用户信息管理")
+@Tag(name = "用户信息管理")
 @RestController
 @RequestMapping("/test/*")
 public class TestController extends BaseController
@@ -30,30 +30,30 @@ public class TestController extends BaseController
         testList.add(new Test("2", "ry", "admin123"));
     }
 
-    @ApiOperation("获取列表")
+    @Operation(summary = "获取列表")
     @GetMapping("list")
     public List<Test> testList()
     {
         return testList;
     }
 
-    @ApiOperation("新增用户")
+    @Operation(summary = "新增用户")
     @PostMapping("save")
     public AjaxResult save(Test test)
     {
         return testList.add(test) ? success() : error();
     }
 
-    @ApiOperation("更新用户")
-    @ApiImplicitParam(name = "Test", value = "单个用户信息", dataType = "Test")
+    @Operation(summary = "更新用户")
+    @Parameter(name = "Test", description = "单个用户信息")
     @PutMapping("update")
     public AjaxResult update(Test test)
     {
         return testList.remove(test) && testList.add(test) ? success() : error();
     }
 
-    @ApiOperation("删除用户")
-    @ApiImplicitParam(name = "Tests", value = "单个用户信息", dataType = "Test")
+    @Operation(summary = "删除用户")
+    @Parameter(name = "Tests", description = "单个用户信息")
     @DeleteMapping("delete")
     public AjaxResult delete(Test test)
     {
